move helper to cmd/interna;
