+++ cmd/osv-scanner/internal/helper/callanalysis_parser.go
