+++ cmd/osv-scanner/internal/helper/callanalysis_parser_test.go
