+++ cmd/osv-scanner/internal/helper/helper.go
